tools/web_direct/pyproject.toml
[project]
name = "web-direct-collector"
version = "0.1.0"
description = "Direct web content collector without API dependencies"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.12.0",
    "click>=8.0.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "trafilatura>=1.6.0",
]

[project.scripts]
web-direct = "tools.web_direct.src.collectors.web_direct.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
